Use a 1-byte DMA buffer for management UART receives.
It sounds silly, but this gives us completion callbacks, so we don't spend
out entire time slice polling the receive buffer for new characters (which
kills performance for tasks that are doing real work). Besides, libcli
wants to process a character at a time, so uart_cli_read just waits on the
completion callback.