"keystore erase" now clears entire keystore flash.
Now that we're using more than just the first few sectors of the
keystore flash, we need a command to clear the whole thing.

This is not quite right yet, because it doesn't yet notify libhal's
ks_flash driver that the entire content of the flash has been yanked
out from under it.

In theory, we should be able to erase the entire flash in a single
operation using the bulk erase command command (0xC7), but I couldn't
get that to do anything (no error reported, no visible effect), so,
for now, we erase by sectors.