Stub out hal_allocate_static_memory() to avoid linker error.
This is a quick fix, so that we can get on with testing the ks9 branch
changes.

A better fix in the long run might be to add a third keystore
("ks_pin_read_only", or some such) which implemented the bare minimum
interface that the bootloader needs and left everything else
unimplemented.  This would require a bit of refactoring the current
PIN code to make it work right with both the bootloader's abbreivated
keystore and the normal token keystore.  Probably worth doing, but a
bit of a can of worms, so postponing for now.