Simplify FMC I/O code.
We're no longer using a preemptive multitasker, and we're on the
Alpha, not the old bridge board, so all we really need to do is read
or write one word at a time via a pointer then wait for the GPIO idle
pin to light.  Well, other than needing to read twice because of a
known hardware problem with the STM32.

So we dont need to fiddle with IRQ, or use the CMSIS SRAM lock, or....